+++ respredai/__init__.py
@@ -3,7 +3,7 @@
 A machine learning pipeline for predicting antimicrobial resistance.
 """
 
-__version__ = "1.9.1"
+__version__ = "1.9.2"
 __author__ = "Ettore Rocchi"
 __email__ = "ettore.rocchi3@unibo.it"